src/tomato_drycal/__init__.py
from datetime import datetime
import serial
import time
from threading import Thread, current_thread, RLock
from tomato.driverinterface_2_0 import Attr, ModelInterface, ModelDevice, Val, Task
from functools import wraps
import pint
import xarray as xr
