pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-git-versioning"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "tomato-drycal"
authors = [
  {name="Peter Kraus", email="peter.kraus@ceramics.tu-berlin.de"},
]

description = "This package is a part of tomato. It has been made for the automation of Jumo Quantrol devices."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyserial >= 3.5",
    "tomato >= 2.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
testing = ["pytest"]

[project.urls]
Repository = "https://github.com/dgbowl/tomato-drycal"
Issues = "https://github.com/dgbowl/tomato-drycal/issues"

[tool.setuptools-git-versioning]
enabled = true
dev_template = "{tag}.dev{ccount}"
dirty_template = "{tag}.dev{ccount}"